requirements.txt
azure-openai>=1.66.0
azure-identity>=1.19.0
azure-search-documents>=11.6.0
azure-keyvault-secrets>=4.9.0
azure-functions>=1.22.0
requests>=2.32.0
flask>=3.1.0
pytest>=8.3.0
azure-devops>=7.0.0
msrest>=0.7.1
pyyaml>=6.0
python-dotenv>=1.0.0
azure-ai-projects>=2.0.0
azure-monitor-opentelemetry>=1.6.0
